+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
     - ``ProjectAppPlugin.update_irods_stats_cache()`` helper (#2213)
     - ``notify_alert_irods_request`` user app setting (#2205)
     - ``UpdateProjectSheetCacheTask`` Taskflow task (#2254)
+    - ``ProjectIrodsFileListAPIView`` ``include_colls`` parameter (#2256)
 - **Taskflowbackend**
     - Project viewer role support (#2200)
     - ``user`` arg for ``TaskflowAPI.submit()`` (#2157)